Refactor type definitions to use instance properties
Class properties and @dataclass are nice enough, but they make
inheritance harder. Being able to call the parent's __init__ cleans up a
lot of things.

Also exposed the IntegerDefintion class, so that we can check if a type
is an integer using `isinstance(..., IntegerDefinition)` (although I'd
like to have a separate property/method for that eventually)