Only list top-level dependencies in the `requirements.txt`
We'll no-longer need to update the `requirements.txt` each time there's
an update in our dependency tree. This seems like a problem with pip in
general... What happens if one of our dependencies changes it's
dependencies, will we be updating and pulling an unneeded package
forever?

We also no longer pull a specific version of Black, its just tooling and
we don't test for it, no need to keep it in sync.