Refactor `Type` interface, disallow nested references
The `Type` interface no-longer remembers if it has been borrowed, this
is now the responsibility of `TypedExpression`. A `Type` is now the same
as would be visible in the user-facing Graphene code.

The `underlying_type` is the type of a `TypedExpression` is the type
that was `typedef`ed in Graphene. A `TypedExpression` may also know the
address of the underlying data (eg. because it was a variable) -- even
though querying the type will say its NOT a reference (the Graphene user
has not borrowed it).

The borrow syntax converts a `TypedExpression` which knows its address
into a Type-system reference. Care must be taken to ensure that any
user-facing expressions return a non-reference type (otherwise the user
would not need to borrow it).