Implement static variables in order to allow `u8[&]` strings
This removes the inbuilt `string` data type, string literals will
instead be initialized a static variable of type `u8[len(string)]`. The
literal itself acts as Borrowed reference to this type.

At the moment there is no syntax to define a static variable of a
non-string type (even tho the interface supports this). All statics are
`private` and must have some initialization (limited to a single line of
llvm-ir).