Convert the `ExpressionTransformer` into an `ExpressionParser`
Before, our parsing of compile time constants was kinda janky... we'd
convert them into a `FlattenedExpression` and then extract the value
when needed. The refactor of the Type system exposed us to a regression
related to this. We now parse root -> leaves so that the leaf nodes
don't need to know that they're in a `Type` vs an `Expression`.