Fix address boundary error caused by large object deallocation
The malloc implementation was bugged. This wasn't caught by the test
because the test struct "happened" to be the exact same size as the other
struct we were mistakenly casting it to.

This was caused by the unsafe casts in the malloc implementation, I want
to rewrite (most) of it using safe language features so this doesn't
happen again. But for now heres a quick fix so the parser starts working.