driver: always compile with `-static`
The ELF loader/program interpreter is tied to the version of `libc`
used, e.g. x86-64 glibc systems use `/lib64/ld-linux-x86-64.so.2` while
musl implementations use `$(syslibdir)/ld-musl-$(ARCH).so.1`.

With the `-nostdlib` flag, the ld linker doesn't set a program
interpreter and produces a statically-linked executable. In
0240dce7587ee76ae58fc40e311db3721173d198, we started linking with LLVM's
lld linker, which deviates from ld and *does* set the program
interpreter, producing a dynamically linked executable. As we are not
using the glibc runtime corresponding to the program interpreter
specified, the program segmentation faults inside the program
interpreter; so before it even starts executing.

We might want to write a custom program interpreter at some point if
want dynamic linking support, but for now, it suffices to always build
static executables.