Remove the hard lark dependency from `graphene_parser`
The lexer-parser is now required to produce a JSON syntax tree and is
responsible for handling its own errors. This will eventually allow us
to plug in the self-hosted parser replacing lark completely.

For now I've emulated the lark data-structures as closely as possible to
minimize algorithmic changes. Later there will be nothing stopping us
from implementing a more appropriate `Transform`/ `Interpret`.

The regressed tests are for syntax errors (which should be handled by
the parser NOT the compiler).