Use a generic parser interface rather than depending on lark
This will allow the self-hosted parser to be easily swapped in: all it
needs to do is conform to the JSON interface (currently based closely on
the lark tree).

In the spirit of removing the lark dependency, I've also reimplemented
the standard lark algorithms we rely on (eg. `Transformer` and
`Interface`). These are a bit awkward and in the future I think it'd be
worth investing time into making everything more ergonomic (which would
also improve the type hinting).