Refactor type and function resolution
I started writing this so that the LSP could dynamically load/ unload
entire translation units worth of symbols but in the refactor it became
convenient to make a number related improvements/ simplifications.

- We no longer differentiate between specialized/ generic functions/ typedefs.
- We now handle function and type resolution with the same code and algorithms.

This has the advantages that: `decltype` is now possible, partial
specializations just need to be turned on, and error messages are improved.