Add memory-map backed file reads (#141)
* Add memory-mapping based io

Instead of streaming reads to a buffer, lets just avoid the extra copy
and let the kernel give us the correct pages. This requires that the
file can be memory mapped and `stat`ed. But the API is MUCH nicer + and
no dynamic allocation is required to read the entire file, the kernel
just gives us whatever pages we want.

I ran a test program that counted digits in a 1GB file (inside a loop),
this saw a 2x speedup (I'm not sure why usr time is higher tho).

Old:
Executed in    4.77 secs    fish           external
   usr time    1.08 secs  190.00 micros    1.08 secs
   sys time    3.68 secs  128.00 micros    3.68 secs

New:
Executed in    2.81 secs    fish           external
   usr time    2.34 secs  189.00 micros    2.34 secs
   sys time    0.46 secs  126.00 micros    0.46 secs

And when I don't loop:

Old:
Executed in  524.10 millis    fish           external
   usr time  114.61 millis    0.00 micros  114.61 millis
   sys time  407.90 millis  298.00 micros  407.60 millis

New:
Executed in  315.17 millis    fish           external
   usr time  263.26 millis  200.00 micros  263.06 millis
   sys time   50.95 millis  133.00 micros   50.82 millis

* Support platform-specific fstat implementations

* Easy PR comments

* Linux kinda jank

* Format with new black